loader: meta stations from build_footpaths to build_stations so they are available during graph building
